Rename package
Переименовала класc DictionaryExeption На exeption
