+++ src/main/java/dictionary/exeption/RemoveException.java
@@ -1,4 +1,4 @@
-package dictionary.DictionaryException;
+package dictionary.exeption;
 
 public class RemoveException extends Exception{
    public RemoveException(String message){ super(message); }
