+++ src/main/java/dictionary/exeption/SearchException.java
@@ -1,4 +1,4 @@
-package dictionary.DictionaryException;
+package dictionary.exeption;
 
 public class SearchException extends Exception{
     public SearchException(String message) {super(message);}
