+++ src/main/java/dictionary/storage/DictionaryStorage.java
@@ -1,7 +1,7 @@
 package dictionary.storage;
 
-import dictionary.DictionaryException.RemoveException;
-import dictionary.DictionaryException.SearchException;
+import dictionary.exeption.RemoveException;
+import dictionary.exeption.SearchException;
 import dictionary.configuration.DictionaryType;
 import dictionary.model.DictionaryLine;
